+++ model-optimizer/automation/package_BOM.txt
@@ -338,6 +338,7 @@ extensions/front/onnx/roialign_ext.py
 extensions/front/onnx/roifeatureextractor_ext.py
 extensions/front/onnx/scatter_ext.py
 extensions/front/onnx/shape_ext.py
+extensions/front/onnx/size_ext.py
 extensions/front/onnx/slice_ext.py
 extensions/front/onnx/softmax_ext.py
 extensions/front/onnx/softmaxONNX_to_softmax.py
@@ -363,6 +364,7 @@ extensions/front/reshape_dim_normalizer.py
 extensions/front/restore_ports.py
 extensions/front/RollWithEmptyAxesReplacer.py
 extensions/front/scatter_normalizer.py
+extensions/front/SizeReplacer.py
 extensions/front/softmax.py
 extensions/front/Softplus_fusion.py
 extensions/front/softsign_replacer.py
@@ -483,7 +485,6 @@ extensions/front/tf/roll_ext.py
 extensions/front/tf/RollRealImagPack.py
 extensions/front/tf/select_ext.py
 extensions/front/tf/sign_ext.py
-extensions/front/tf/SizeReplacer.py
 extensions/front/tf/slice_ext.py
 extensions/front/tf/softmax_ext.py
 extensions/front/tf/softplus_ext.py
