+++ model-optimizer/unit_tests/extensions/front/size_replacer_test.py
@@ -6,7 +6,7 @@ import unittest
 import numpy as np
 from generator import generator, generate
 
-from extensions.front.tf.SizeReplacer import SizeFrontReplacer
+from extensions.front.SizeReplacer import SizeFrontReplacer
 from mo.front.common.partial_infer.utils import int64_array
 from mo.utils.ir_engine.compare_graphs import compare_graphs
 from unit_tests.utils.graph import build_graph, regular_op_with_empty_data, result, connect, \
